<- organise

+++ hooks/clean.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+_CHOICE=$1
 CURRENT=`pwd`
 echo "================= CLEAN SCRIPT ==================="
 tempsDirectory="temps/"
@@ -8,12 +9,14 @@ if [ -d $CURRENT/$tempsDirectory ]; then
 else
 	echo "$CURRENT/$tempsDirectory does not exist !"
 fi
-executeDiretory="execute/"
-if [ -d $CURRENT/$executeDiretory ]; then
-	echo "REMOVING $CURRENT/$executeDiretory"
-	rm -r $CURRENT/$executeDiretory
-else
-	echo "$CURRENT/$executeDiretory does not exist !"
+if [ $_CHOICE -eq 1 ]; then
+	executeDiretory="execute/"
+	if [ -d $CURRENT/$executeDiretory ]; then
+		echo "REMOVING $CURRENT/$executeDiretory"
+		rm -r $CURRENT/$executeDiretory
+	else
+		echo "$CURRENT/$executeDiretory does not exist !"
+	fi
 fi
 outputDiretory="output/"
 if [ -d $CURRENT/$outputDiretory ]; then
